Re-factor StaticMetadataValue into FixedMetadataValue
It was requested that instead of using StaticMetadataValue as a new
class, that instead we'd fix FixedMetadataValue to get most of the
performance improvements, even if it came at a slight cost in memory
over the clean implementation.
