+++ src/main/java/org/bukkit/metadata/FixedMetadataValue.java
@@ -6,9 +6,15 @@ import java.util.concurrent.Callable;
 
 /**
  * A FixedMetadataValue is a special case metadata item that contains the same value forever after initialization.
- * Invalidating a FixedMetadataValue has no affect.
+ * Invalidating a FixedMetadataValue has no effect.
+ *
+ * This class extends LazyMetadataValue for historical reasons, even though it overrides all the implementation
+ * methods. it is possible that in the future that the inheritance hierarchy may change.
  */
 public class FixedMetadataValue extends LazyMetadataValue {
+    /** Store the internal value that is represented by this fixed value. */
+    private final Object internalValue;
+
     /**
      * Initializes a FixedMetadataValue with an Object
      *
@@ -16,10 +22,17 @@ public class FixedMetadataValue extends LazyMetadataValue {
      * @param value the value assigned to this metadata value.
      */
     public FixedMetadataValue(Plugin owningPlugin, final Object value) {
-        super(owningPlugin, CacheStrategy.CACHE_ETERNALLY, new Callable<Object>() {
-            public Object call() throws Exception {
-                return value;
-            }
-        });
+        super(owningPlugin);
+        this.internalValue = value;
+    }
+
+    @Override
+    public void invalidate() {
+
+    }
+
+    @Override
+    public Object value() {
+        return internalValue;
     }
 }
