+++ src/main/java/org/bukkit/metadata/LazyMetadataValue.java
@@ -16,7 +16,7 @@ import org.bukkit.plugin.Plugin;
 public class LazyMetadataValue extends MetadataValueAdapter implements MetadataValue {
     private Callable<Object> lazyValue;
     private CacheStrategy cacheStrategy;
-    private SoftReference<Object> internalValue = new SoftReference<Object>(null);
+    private SoftReference<Object> internalValue;
     private static final Object ACTUALLY_NULL = new Object();
 
     /**
@@ -40,9 +40,14 @@ public class LazyMetadataValue extends MetadataValueAdapter implements MetadataV
         super(owningPlugin);
         Validate.notNull(cacheStrategy, "cacheStrategy cannot be null");
         Validate.notNull(lazyValue, "lazyValue cannot be null");
-
+        this.internalValue = new SoftReference<Object>(null);
         this.lazyValue = lazyValue;
         this.cacheStrategy = cacheStrategy;
+    }
+
+    /** Protected special constructor used by FixedMetadataValue to bypass standard setup. */
+    protected LazyMetadataValue(Plugin owningPlugin) {
+        super(owningPlugin);
     }
 
     public Object value() {
